Setup to run a big collect using all the basic robust cost models in gtsam.  Also keep track of timing for each run.
Noticed when I was running things that some cost functions seemed to be taking must longer to converge, so added timing.  Save the "options" arrays as part of a file now too so only one file is saved for all the results, rather than having multiple files, one for each permutation.